backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic-settings==2.6.1
paho-mqtt==2.1.0
alembic==1.14.0
python-dotenv==1.0.1
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
email-validator==2.2.0
